feature: customer cart and its functions
- added customer cart and navigation to it by customers.
- added to cart and remove from cart functionality.

+++ database/migrations/2024_06_11_094317_create_products_table.php
@@ -17,8 +17,8 @@ return new class extends Migration
             $table->foreignId('product_category_id')->nullable()->references('id')->on('product_categories')->cascadeOnDelete()->cascadeOnUpdate();
             $table->string('name', 128);
             $table->string('description', 256)->nullable();
-            $table->decimal('price', 10)->default(0);
-            $table->decimal('discounted_price', 10)->default(0);
+            $table->float('price', 10)->default(0);
+            $table->float('discounted_price', 10)->default(0);
             $table->integer('quantity')->default(0);
             $table->boolean('is_active')->default(true);
             $table->timestamps();
